fibonacci4

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,9 +6,9 @@
 
 int main(void)
 {
-	unsigned long int fib1 = 1;
-	unsigned long int fib2 = 2;
-	unsigned long int fib3;
+	unsigned long long int fib1 = 1;
+	unsigned long long int fib2 = 2;
+	unsigned long long int fib3;
 	int i;
 
 	printf("%lu, %lu, ", fib1, fib2);
